requirements.txt
streamlit
PyPDF2
newspaper3k
openai
gTTS
